requirements.txt
setuptools~=46.0.0
requests~=2.23.0
lxml~=4.5.0
notifyme~=0.2.0
beautifulsoup4~=4.8.2
PyYAML~=5.3.1